Introduce interface for overlapping output detection
This should make testing simpler.
