+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/DefaultTaskCacheabilityResolverTest.groovy
@@ -25,7 +25,7 @@ import org.gradle.api.specs.Spec
 import org.gradle.internal.execution.caching.CachingDisabledReason
 import org.gradle.internal.execution.caching.CachingDisabledReasonCategory
 import org.gradle.internal.file.RelativeFilePathResolver
-import org.gradle.internal.fingerprint.OverlappingOutputs
+import org.gradle.internal.fingerprint.overlap.OverlappingOutputs
 import spock.lang.Specification
 
 import javax.annotation.Nullable
