+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/ResolveBeforeExecutionOutputsTaskExecuterTest.groovy
@@ -26,8 +26,8 @@ import org.gradle.api.internal.tasks.TaskStateInternal
 import org.gradle.api.internal.tasks.properties.OutputFilePropertySpec
 import org.gradle.api.internal.tasks.properties.TaskProperties
 import org.gradle.internal.execution.history.AfterPreviousExecutionState
-import org.gradle.internal.fingerprint.OverlappingOutputs
 import org.gradle.internal.fingerprint.impl.AbsolutePathFingerprintingStrategy
+import org.gradle.internal.fingerprint.overlap.OverlappingOutputs
 import org.gradle.internal.snapshot.FileSystemSnapshot
 import spock.lang.Specification
